+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # ⏩ vNext
+- Add support for custom transformers on parsed values
 
 # v4.2.2
 - Fix an issue where the new 53-bit oversize error is emitted when the `float` format is selected, even though it is